TST: Update and convert tests to unittest
